assignment_4

+++ src/assignment2/Task_5.java
@@ -22,7 +22,7 @@ public class Task_5 {
         */
         Scanner sc=new Scanner(System.in);
         System.out.println("Enter the radius of a circle:");
-        double radius=sc.nextInt();
+        double radius=sc.nextDouble();
         double pi=3.14159;
         double circumference=2*pi*radius;
         double area=pi*(radius*radius);
